fix: typo

+++ cpp/tensorrt_llm/executor/cache_transmission/mooncake_utils/transferAgent.h
@@ -20,6 +20,8 @@
 #include <atomic>
 #include <mutex>
 #include <thread>
+#include <unordered_set>
+#include <vector>
 
 #include "tensorrt_llm/executor/transferAgent.h"
 #include "transfer_engine_c.h"
@@ -119,7 +121,7 @@ private:
 
     mutable std::mutex mMutex;
     transfer_engine_t mEngine;
-    std::unordered_map<uint64_t, std::shared_ptr<MemoryInfo>> mMemRegInfo;
+    std::unordered_map<uintptr_t, std::shared_ptr<MemoryInfo>> mMemRegInfo;
     std::string local_agent_name_;
     std::string segment_name_;
 };
